requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt ./requirements.in
#
asgiref==3.7.2
    # via django
bleach[css]==6.0.0
    # via django-markdownify
caldav==1.3.6
    # via -r ./requirements.in
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.2.0
    # via requests
cryptography==41.0.3
    # via
    #   django-encrypted-model-fields
    #   pyjwt
defusedxml==0.7.1
    # via python3-openid
django==4.0.10
    # via
    #   -r ./requirements.in
    #   django-allauth
    #   django-encrypted-model-fields
    #   django-markdownify
    #   django-simple-csp
    #   django-token-bucket
    #   djangorestframework
django-allauth==0.55.0
    # via -r ./requirements.in
django-encrypted-model-fields==0.6.5
    # via -r ./requirements.in
django-friendly-tag-loader==1.3.1
    # via -r ./requirements.in
django-markdownify==0.9.3
    # via -r ./requirements.in
django-pipeline==2.1.0
    # via -r ./requirements.in
django-simple-csp==0.3.dev1
    # via -r ./requirements.in
django-token-bucket==0.2.4
    # via -r ./requirements.in
django-widget-tweaks==1.4.12
    # via -r ./requirements.in
djangorestframework==3.14.0
    # via -r ./requirements.in
icalendar==5.0.7
    # via
    #   -r ./requirements.in
    #   caldav
    #   recurring-ical-events
    #   x-wr-timezone
idna==3.4
    # via requests
libsass==0.22.0
    # via libsasscompiler
libsasscompiler==0.1.9
    # via -r ./requirements.in
lxml==4.9.3
    # via caldav
markdown==3.4.4
    # via django-markdownify
mysqlclient==2.2.0
    # via -r ./requirements.in
oauthlib==3.2.2
    # via requests-oauthlib
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via django-allauth
python-dateutil==2.8.2
    # via
    #   icalendar
    #   recurring-ical-events
    #   vobject
python3-openid==3.2.0
    # via django-allauth
pytz==2023.3
    # via
    #   -r ./requirements.in
    #   caldav
    #   djangorestframework
    #   icalendar
    #   recurring-ical-events
    #   x-wr-timezone
recurring-ical-events==2.0.2
    # via caldav
requests==2.31.0
    # via
    #   caldav
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
six==1.16.0
    # via
    #   bleach
    #   python-dateutil
sqlparse==0.4.4
    # via django
tinycss2==1.1.1
    # via bleach
tzlocal==5.0.1
    # via caldav
urllib3==2.0.4
    # via requests
uwsgi==2.0.22
    # via -r ./requirements.in
vobject==0.9.6.1
    # via caldav
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
x-wr-timezone==0.0.5
    # via recurring-ical-events
